aplit auth function api, and add more concurrency
