+++ scripts/split-e2e-tests.ts
@@ -3,8 +3,20 @@ import * as glob from 'glob';
 import { join } from 'path';
 import * as fs from 'fs-extra';
 
-const CONCURRENCY = 12;
-const AWS_REGIONS_TO_RUN_TESTS = ['us-east-2', 'eu-west-1', 'ap-southeast-1', 'ap-south-1', 'ap-southeast-2'];
+const CONCURRENCY = 30;
+const AWS_REGIONS_TO_RUN_TESTS = [
+  'us-east-1', 
+  'us-east-2', 
+  'us-west-2', 
+  'eu-west-1', 
+  'eu-west-2', 
+  'eu-central-1', 
+  'ap-northeast-1', 
+  'ap-northeast-2', 
+  'ap-southeast-1', 
+  'ap-southeast-2', 
+  'ap-south-1'
+];
 
 // This array needs to be update periodically when new tests suites get added
 // or when a test suite changes drastically
